[Chore] Move SidebarToggle render prop to its own file
