+++ packages/react-search-ui-views/src/layouts/Layout.js
@@ -1,6 +1,7 @@
 import PropTypes from "prop-types";
 import React from "react";
 
+import SidebarToggle from "./SidebarToggle";
 import { appendClassName } from "../view-helpers";
 
 function Layout({
@@ -54,40 +55,5 @@ Layout.propTypes = {
   bodyHeader: PropTypes.node,
   sideContent: PropTypes.node
 };
-
-class SidebarToggle extends React.Component {
-  static propTypes = { children: PropTypes.func };
-
-  state = { isSidebarToggled: false };
-
-  toggleSidebar = () => {
-    this.setState(({ isSidebarToggled }) => ({
-      isSidebarToggled: !isSidebarToggled
-    }));
-  };
-
-  render() {
-    const { isSidebarToggled } = this.state;
-
-    const renderToggleButton = label => (
-      <button
-        hidden
-        type="button"
-        className="sui-layout-sidebar-toggle"
-        onClick={this.toggleSidebar}
-      >
-        {label}
-      </button>
-    );
-
-    const renderToggleClass = className =>
-      appendClassName(
-        className,
-        isSidebarToggled ? `${className}--toggled` : null
-      );
-
-    return this.props.children({ renderToggleButton, renderToggleClass });
-  }
-}
 
 export default Layout;
